factored out the new add survey modal
